Add some logic to review/enrollment

+++ src/main/java/com/task/onlinecoursemanagementsystem/student/enrollment/controller/dto/ProgressUpdateDto.java
@@ -4,10 +4,12 @@ import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record ProgressUpdateDto(
         @NotNull
         @DecimalMin(value = "0.0")
         @DecimalMax(value = "100.0")
-        Double newProgress
+        BigDecimal newProgress
 ) {
 }
